# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: CtxOS API
  description: Context-driven OS Intelligence and Analysis API
  version: 1.0.0
  contact:
    name: CtxOS Team
    email: api@ctxos.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.ctxos.com
    description: Production server
...
paths:
  /api/v1/auth/login:
    post:
      tags: [auth]
      summary: Authenticate user
      description: Authenticate user and return access and refresh tokens
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token
      description: Use refresh token to get new access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RefreshResponse'
        '401':
          description: Invalid or expired refresh token
  /api/v1/auth/me:
    get:
      tags: [auth]
      summary: Get current user info
      description: Get information about the currently authenticated user
      operationId: getCurrentUser
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User information retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserInfo'
        '401':
          description: Unauthorized
  /api/v1/auth/logout:
    post:
      tags: [auth]
      summary: Logout user
      description: Logout and revoke the current token
      operationId: logout
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
  /api/v1/auth/verify:
    get:
      tags: [auth]
      summary: Verify token
      description: Verify if token is valid and return token info
      operationId: verifyToken
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Token is valid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenVerification'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          description: User username
        password:
          type: string
          description: User password
    LoginResponse:
      type: object
      required: [access_token, refresh_token, token_type, expires_in, user_info]
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: JWT refresh token
        token_type:
          type: string
          enum: [bearer]
          description: Token type
        expires_in:
          type: integer
          description: Token expiration time in seconds
        user_info:
          $ref: '#/components/schemas/UserInfo'
    RefreshRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
          description: Refresh token
    RefreshResponse:
      type: object
      required: [access_token, token_type, expires_in]
      properties:
        access_token:
          type: string
          description: New access token
        token_type:
          type: string
          enum: [bearer]
        expires_in:
          type: integer
          description: Token expiration time in seconds
    UserInfo:
      type: object
      required: [user_id, username, role, permissions]
      properties:
        user_id:
          type: string
          description: User ID
        username:
          type: string
          description: Username
        email:
          type: string
          description: User email
        role:
          type: string
          enum: [admin, analyst, viewer, api_client]
          description: User role
        permissions:
          type: array
          items:
            type: string
          description: User permissions
        is_active:
          type: boolean
          description: Whether user is active
        created_at:
          type: string
          format: date-time
          description: Account creation time
        last_login:
          type: string
          format: date-time
          description: Last login time
    TokenVerification:
      type: object
      required: [valid, token_info, expires_in]
      properties:
        valid:
          type: boolean
          description: Whether token is valid
        token_info:
          type: object
          description: Token information
        expires_in:
          type: integer
          description: Time until token expires in seconds

    # Common Schemas
    StatusResponse:
      type: object
      required: [status, message]
      properties:
        status:
          type: string
          enum: [success, error]
          description: Operation status
        message:
          type: string
          description: Status message
        details:
          type: object
          description: Additional details
    ErrorResponse:
      type: object
      required: [error, status_code]
      properties:
        error:
          type: string
          description: Error description
        status_code:
          type: integer
          description: HTTP status code
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
...